+++ CHANGELOG.md
@@ -2,9 +2,10 @@
 
 ## [1.4.2]
 - The servers in the Discord chat navigation tree are collapsed by default
+- Custom emojis support
 
 ## [1.4.1]
-- small fix
+- Small fix
 
 ## [1.4.0]
 - Discord chat [BETA]